Return actual error instead of generic error message (if possible)
This means that the actual error message is passed to the observer. The
observer can e.x. show the error message to the user.
Concrete example: If you already have a home-entity and try to set
another entity to ‘ home’, then the server returns an appropriate error
message that should be shown to the user